feat: Improve `ContextMenu` component for `accessibility`
